Add station data to system imports/exports
Uses a new connection that attaches multiple databases.

This allows flexibility to manage them as three different datastores but can still do cross joins.

SQLlite supports attaching 125 databases in this way, which is more than enough.

I'm taking a modest approach to this to see if this surfaces any issues or if I should normalize the data in the trade db.

If it performs well over time for queries, I might remove some of the station information from the tradedb (such as station name and system information)