Add support for querying by System Address as well as by System Name
This is primarily in the service of explicitly improving support for disambiguation amoung systems with the same name, which is an edge case in the game, but by may increasingly become an issue with colonisation as there can now be multiple populated systems with the same name.

* All system endpoints can now use /system/address/:systemAddress anywhere /system/name/:systemName is supported.
* All commodity endpoints can now use systemAddress in place of systemName (if both are specified, systemAddress is preferred)
* New market endpoint to allow querying by Market ID specifically (as well as system + station name). As Market IDs are unique, no system name or system address is required.
* When querying for a system, if it's one of the > 1000 duplicates in the database, a list of the matches are also returned in an array titled 'disambiguation'. This contains typically only 1 other entry, but in cases can be 2 or even 3 other systems.
* When searching for a system by name, if there are other systems with the same name, the flag 'ambiguous' is returned. Explicitly fetching the system by name (or by address) will explictly return the other matches for that result.
* Improved handling of disambiguation of results in commodity queries.

All of the changes are additive to responses, none of the changes are intended to be breaking, execept with respect to a minor change in the response to the undocumented / experimental system status endpoint.

Queries for system names will continue to work as before, and the API wil automatically try to match base on case sensitivity of any supplied system name, but will still support returning data for one of the matchign systems if there is a case-insensitive match. The automated match should always resolve to the same system.

The user interface for the website will continue to prefer names over addresses in user visible URLs and to highlight where there is known to be ambiguity, but is being updated to allow a system address to be specified in all cases where a system name can be specified.

These API changes are not yet documented but the intent is that these new features will be and that are now useable and supported.